refactor: extract ResourceFeatureItem component
